+++ changelog.md
@@ -1,3 +1,7 @@
+# 1.0.1
+- Fixed RobTop's bug that completely freaks out cam (thanks to @Brittank for telling me about this bug)
+- Added the keybind for quick access to settings (default is `F7`)
+
 # 1.0.0
 - Static position
   - Panning during playtest with mouse `middle click`